+++ src/lib/pages/proposal-details/components/proposal-overview/status-summary/SummaryStatusTime.tsx
@@ -1,4 +1,5 @@
 import { Text } from "@chakra-ui/react";
+import { isNull } from "lodash";
 
 import type { ProposalData } from "lib/types";
 import { ProposalStatus } from "lib/types";
@@ -34,7 +35,7 @@ export const SummaryStatusTime = ({ proposalData }: StatusTimeProps) => {
     return (
       <Text variant="body2" textAlign={{ base: "end", md: "start" }}>
         Voting ends in{" "}
-        {proposalData.votingEndTime ? (
+        {!isNull(proposalData.votingEndTime) ? (
           <Countdown endTime={proposalData.votingEndTime} isString={false} />
         ) : (
           "N/A"
@@ -50,7 +51,7 @@ export const SummaryStatusTime = ({ proposalData }: StatusTimeProps) => {
     >
       {getResolvedPrefix(proposalData.status)}
       {" at "}
-      {proposalData.resolvedTimestamp
+      {!isNull(proposalData.resolvedTimestamp)
         ? formatUTC(proposalData.resolvedTimestamp)
         : "N/A"}
     </Text>