+++ src/lib/pages/proposal-details/components/vote-details/VoteDetailsAccordionItem.tsx
@@ -9,16 +9,27 @@ import type { ReactNode } from "react";
 import { ProposalStepper } from "../proposal-stepper";
 import type { ProposalData } from "lib/types";
 
+interface VoteDetailsAccordionItemProps {
+  step: number;
+  proposalData: ProposalData;
+  children: ReactNode;
+  onClick: () => void;
+  isDisabled?: boolean;
+}
+
 export const VoteDetailsAccordionItem = ({
   step,
   proposalData,
   children,
-}: {
-  step: number;
-  proposalData: ProposalData;
-  children: ReactNode;
-}) => (
-  <AccordionItem borderTop="1px solid" borderColor="gray.700">
+  onClick,
+  isDisabled,
+}: VoteDetailsAccordionItemProps) => (
+  <AccordionItem
+    borderTop="1px solid"
+    borderColor="gray.700"
+    onClick={onClick}
+    isDisabled={isDisabled}
+  >
     <AccordionButton py={3} px={0}>
       <ProposalStepper step={step} proposalData={proposalData} />
       <AccordionIcon color="gray.600" ml="auto" />