+++ src/lib/pages/proposal-details/components/vote-details/voting-period/VotingQuorum.tsx
@@ -19,7 +19,7 @@ export const VotingQuorum = ({
   params,
   votesInfo,
   isLoading,
-}: VoteDetailsProps) => {
+}: Omit<VoteDetailsProps, "period">) => {
   const isMobile = useMobile();
   if (isLoading) return <Loading my={0} />;
   if (!params || !votesInfo) return <ErrorFetchingProposalInfos />;